Mention if a physical attac didn't cause damage

+++ src/player.c
@@ -2046,13 +2046,13 @@ void player_damage_take(player *p, damage *dam, player_cod cause_type, int cause
     switch (dam->type)
     {
     case DAM_PHYSICAL:
-        if (dam->amount > player_get_ac(p))
+        if (dam->amount > (gint)player_get_ac(p))
         {
             dam->amount -= player_get_ac(p);
 
-            if (dam->amount >= 8 && dam->amount >= p->hp_max/4)
+            if (dam->amount >= 8 && dam->amount >= (gint)p->hp_max/4)
                 log_add_entry(nlarn->log, "Ouch, that REALLY hurt!");
-            else if (dam->amount >= p->hp_max/10)
+            else if (dam->amount >= (gint)p->hp_max/10)
                 log_add_entry(nlarn->log, "Ouch!");
 
             player_hp_lose(p, dam->amount, cause_type, cause);
@@ -2064,13 +2064,13 @@ void player_damage_take(player *p, damage *dam, player_cod cause_type, int cause
         break;
 
     case DAM_MAGICAL:
-        if (dam->amount > (guint)player_effect(p, ET_RESIST_MAGIC))
+        if (dam->amount > (gint)(guint)player_effect(p, ET_RESIST_MAGIC))
         {
             dam->amount -= player_effect(p, ET_RESIST_MAGIC);
 
-            if (dam->amount >= 8 && dam->amount >= p->hp_max/4)
+            if (dam->amount >= 8 && dam->amount >= (gint)p->hp_max/4)
                 log_add_entry(nlarn->log, "Ouch, that REALLY hurt!");
-            else if (dam->amount >= p->hp_max/10)
+            else if (dam->amount >= (gint)p->hp_max/10)
                 log_add_entry(nlarn->log, "Ouch!");
 
             player_hp_lose(p, dam->amount, cause_type, cause);
@@ -2083,7 +2083,7 @@ void player_damage_take(player *p, damage *dam, player_cod cause_type, int cause
         break;
 
     case DAM_FIRE:
-        if (dam->amount > (guint)player_effect(p, ET_RESIST_FIRE))
+        if (dam->amount > (gint)player_effect(p, ET_RESIST_FIRE))
         {
             dam->amount -= player_effect(p, ET_RESIST_FIRE);
 
@@ -2097,7 +2097,7 @@ void player_damage_take(player *p, damage *dam, player_cod cause_type, int cause
         break;
 
     case DAM_COLD:
-        if (dam->amount > (guint)player_effect(p, ET_RESIST_COLD))
+        if (dam->amount > (gint)player_effect(p, ET_RESIST_COLD))
         {
             dam->amount -= player_effect(p, ET_RESIST_COLD);
